Add requirements.txt with hashes for dependencies
Add a bash script and necessary helper files to generate the
requirements.txt from requirements.in. This improves security as pip
install will verify the hashes.

Though it can be periodically manually regenerated by running the shell
script, requirements.txt should be static rather than generated as part
of the build process for the Docker image containing TFF. This way if
an attacker gains control over the PyPI account of a library, the new
hash on PyPI will not match the one in requirements.txt.

Change-Id: I9a58d20f8957c45b301769b5bd96be9463f2bfbc