pipeline_transforms support for encryption.
The `RecordDecoder` and `RecordEncoder` structs can be used to parse and
produce either plaintext or encrypted inputs and outputs. The
unencrypted mode is useful during development, but can be disabled when
a `RecordDecoder` is constructed to guarantee that an app doesn't allow
unencrypted inputs (if that's important).

The example apps currently only produce unencrypted results; they'll be
updated to support encryption separately.

Note that the proto changes are safe because messages of these types
have not been serialized yet.

Change-Id: Iba98161fb562377a8354de69d10ec85c19fb97f9