Add example app support for encrypted records.
A new `Initialize` method passes optional configuration to the service
and returns the public key to use for encrypting inputs. This change
renumbers the method numbers since initialization is logically first;
this is okay because there are no active callers of these methods.

The sum service has been updated to accept encrypted inputs; it always
returns unencrypted outputs. The square service has been updated to
return an encrypted output iff the input was encrypted.

Bug: #287284320
Change-Id: I9939a384f31ebb277b3745251c47bfc3749f6d4a