Make the repo buildable with bazelisk.
Replacing `@rules_proto_grpc` with direct uses of the rules provided by
`@com_github_grpc_grpc` allows the FCP rules to be reused and fixes some
build errors when using bazelisk without a special Docker container.

The specific C++ toolchain to use is not specified, which means that
this build isn't hermetic (yet); that can be addressed separately if we
find a good solution.

Also make a few minor C++ style fixes.

Bug: #291296933
Change-Id: I6a9ba4e941b82aa5fc0459933038ebcd44aac2e8