Initial Ledger service implementation.
This initial implementation verifies that requests match the
DataAccessPolicy and rewraps keys, but it does not (a) attest the
requesting application, (b) track usage budgets, or (c) have any form of
replication. Missing functionality will be added in subsequent changes.

Note that the BudgetTracker interface is likely to change a little bit
(e.g., to avoid recomputing some derived state about the access policy);
it's provided as-is now to demonstrate the general shape of interactions
with a budget tracking module.

Bug: #288282266
Change-Id: I5e85aac792542e089c5464c7dfc06c990145531a