Implement Transform to execute TFF aggregations.
The ConfigureAndAttest and Transform methods are implemented so that configuration can be provided to execute an aggregation on unencrypted client inputs. The aggregation is specified as a serialized TFF computation, and the client inputs are expected to be serialized TFF values.

Implementing execution of a client_work computation will be in a follow-up change.

Verified by running:
$ bazelisk test tff_worker/server:all
$ docker build -f tff_worker/server/Dockerfile -t pipeline_transform_server .

Change-Id: Ica330e453748bdc586bc113efc2ea94605eb2834