Install deps for building TensorFlow & update FCP.
The latest version of the tff_worker_configuration_py_proto takes a
dependency on TensorFlow protos, so the WORKSPACE file and .bazelrc need
to be updated to install necessary TensorFlow dependencies.

This will also be useful in the future for building TFF from source for
a C++ version of the server.

Also fix a missing dependency issue that surfaces when running the
server as a standalone binary.

Verified:
```
> bazelisk test //tff_worker/server:all
> docker build -f tff_worker/server/Dockerfile -t pipeline_transform_server .
> docker run -i -p 127.0.0.1:50051:50051 pipeline_transform_server:latest &
> bazel-bin/tff_worker/client/pipeline_transform_cpp_client
```

Change-Id: I434e15a2edb4761c47392cbe11520e1a82a9d65e