Implement client work transform using TFF.
The client input format specification in the configuration proto is used
to determine how to parse the bytes. Currently only a TensorFlow
checkpoint in columnar format (as produced by a Brella FedSQL query) is
supported.

An OrderedDict of tensors is created from the parsed checkpoint. The
client_work TensorFlow Federated computation is checked to ensure it is
compatible with the input. Then the computation is executed on the client input dictionary and the broadcasted data (provided in the
configuration) and finally the output is serialized as a tff
`executor_pb2.Value` proto and returned in the response.

Change-Id: Id960ee686c4300bd3063f9f2e4d6b83c5444cfbc