Expect clients-placed inputs to client_work comp.
The client_work computation generated by DistributeAggregateForm will
expect client inputs and broadcast inputs to be placed @CLIENTS with all_equal set to False. Given that in Trusted Brella the client work computation is executed one client at a time and we construct the types to ensure that the checkpoint is compatible with the computation's expected input, this requires some modifications to the code to ensure the types will be compatible and the computation can be executed properly.

This also requires updating the tests to use tff.federated_computation which is a computation that includes placements.

BUG: 289583377
Change-Id: I33cc7edd5fe62072317e0be0a1384a7c1b86091a