Add creation of container OCI runtime bundle to CI
Modify the kokoro bazel build script so that release builds also build the docker container and create an OCI runtime bundle for use with Oak Containers, using scripts modeled off of the Oak repo example container.

The bazelisk build now runs outside the Docker build process and artifacts are copied in to the build context, which should speed up the local development process and CI by allowing for more caching by Bazel.

Building the docker container is not added to presubmit builds as it
adds 10 minutes to presubmit. Now that the bazel build process happens
outside the container the chances that a code change will break the
docker build without breaking the bazel test process that runs on
presubmit are low.

BUG: 291999517
Change-Id: I81c49391770f9946d8da70d05c789860cf92910b