Notify Oak orchestrator of container startup.
The guest must notify the host that it is ready to start serving
requests by calling the Oak Containers Orchestrator API.

This also works around some gaps in Oak Containers. There are no Python
targets for the necessary APIs, so we patch them in, and patch the
dependencies to remove Java targets to avoid needing to take any
dependencies on Java infrastructure in this repo. Additionally, we
modify the /etc/hosts file in the container to include a mapping for
localhost to work around a name resolution issue which manifests when
TFF creates a local GRPC channel to invoke C++ code from Python.

BUG: 291999517
Change-Id: Id689a21480420b41a7a2ffb3a05ddd77cc6c5431