Create dev container with necessary deps.
Use this development docker container for performing local and Kokoro
builds. This speeds up iteration on the code as the expensive build deps
only need to be downloaded once to create the dev container, while
ensuring the build is reproducible on different developers machines.

The next step is to setup uploads of the dev container to Google Cloud.

BUG: 305298267
Change-Id: I6fe20369f75c251f7fa88bd44a4d12c9b72a0066