Implement decryption for container inputs.
The RecordDecryptor class will be used by C++ containers to decrypt a
record that is rewrapped by the Ledger for decryption by
this particular container.

In order for the Ledger to rewrap a message so it can be used exactly
once by this container, the class also provides a method to get the public
key for the container, and a signature of that public key as well as the container configuration with a key rooted in the attestation of this
container. These will be provided to the Ledger so the Ledger can verify
the identity of the container before using the public key to rewrap the
data. Additionally, a method to generate a nonce is provided. The Ledger
will bind the nonce to the rewrapped data as associated data, and the
container will save the nonce and use it exactly once to decrypt the
data. This ensures that a malicious server operator cannot replay
messages to be used multiple times by the same container.

BUG: 294292464

Change-Id: Ia1baa186aba2d7a18b0acf8ed02004bf26cda792