Fix toplevel build script to work locally or on CI.
Previously the addition of exporting the XDG_RUNTIME_DIR variable broke
local builds but was necessary to get CI release builds working.

This change adds a way to default to the local value of
XDG_RUNTIME_DIR when running the script locally.

For simplicity when reading the script, Cargo specific steps are now
split into their own script which is invoked from the toplevel script.

BUG: 318572462

Change-Id: I33ab742a15b4530a939d5252be6a9baa4770ef93