Create a new SQLite db connection for each `Transform`.
Our microbenchmarks did not find a statistically significant performance
difference for this change. Given that the performance is similar, we
want to make this change since it's safer to have separate databases
to prevent a malicious query from saving data and copying it into a
new table in the same database that would outlive the query.

Change-Id: I9c089484703d9b5169595a5782d76d427dca42ca