Build C++ OCI runtime bundles with bazel.
Bundles built this way are smaller (165 MB -> 35 MB), faster to build,
and reproducible -- and the build system is simpler. The tff_worker
container has not been converted because it requires a significantly
more complex base image.

Also tag the pipeline_transform_server_benchmarks test as manual so it
doesn't need to be specifically excluded in bazel_build_with_deps.sh.

Change-Id: I4fb6e3d153b48afbf0757aedcf2a06479aca4c39