Switch toolchain from clang to gcc.
The clang toolchain doesn't provide a sysroot, so it uses some libraries
and headers from the local system. The gcc toolchain does not have this
limitation.

The gcc toolchain also doesn't link with `-ldl` by default, which causes
the sqlite build to break. It's necessary to patch that build to
properly list library dependencies

Finally, add some options to .bazelrc to make builds more hermetic and
make sure we aren't accidentally relying on a local toolchain.

Bug: 322011540
Change-Id: Id795625ef662c5f9b6ad9ba79e80a3d7b7a97f72