Move test TFF computations to "testing/" and add serialized versions.
This change prepares the testing computations to be used in both Python
and C++ TFF containers in three ways:

1) Moves the Python declarations of the TFF computations into the
   testing/ directory.
2) Creates a binary for generating serialized versions of the test TFF
   computations.
3) Uses the binary from (2) to generate serialized versions of the test
   TFF computations and adds those to the repo.

Parts (2) and (3) are pre-work for creating the C++ version of the
container (so tests can be easily added).

Additionally, this change fixes some style issues in
checkpoint_test_utils.py and pipeline_transform_server_test.py.

Change-Id: Ie5ccdb1ddf9c8bf27300b716b34483657d15f9c1