Add sanitizer support to the cfc repo.
The gcc toolchain we use for build reproducability does not make it easy
to enable sanitizers to run on tests, so add a clang toolchain for use
with sanitizers, and configure the .bazelrc with the options for each
sanitizer.

Update the build script to support running the sanitizers in
addition to the regular tests when the argument provided to the script
is "continuous". This will enable us to configure CI to run tests with
sanitizers.

Fix a small undefined behavior issue in the sqlite_adapter_test detected
by ubsan: the pointer to status().message is null when the status is OK,
so a nullptr was being passed to the LOG implementation.

Tested locally by running `./scripts/build.sh continuous" and verifying
that tests were run with the different sanitizers and passed. Also
verified that when I introduced a use=after-free bug in a test and ran
the test with --config=asan, the test failed with the relevant error
message.

Bug: 324146396
Change-Id: I1c1ce081e1d16da6da4db916f10153f7aed38de5