Verify attestation during AuthorizeAccessRequests.
Verification is only performed if the access policy includes
ReferenceValues; this ensures that existing use-cases continue to work
while we finish plumbing through Evidence and Endorsements.

Note that I haven't been able to find test
Evidence/Endorsements/ReferenceValues that pass validation yet since the
examples in the Oak repo (a) don't have application signing keys that
can be used to sign additional data and (b) aren't exposed by any of the
Oak crates, so we'd need to import the Oak repo as a git submodule.

Bug: 288331695
Change-Id: Ic4bdacfc0c81eec244cea2568690aa38e17c9cf2