Implement Transform API for client work for the TFF-based TransformService.
This change adds the Transform API to the TFF-based C++
TransformService. This requires setting up a TFF execution stack inside
the service. Currently, this only enables this API for client work;
server-side aggregation transforms will be added in a follow-up change.

Additional changes:

 * The tff_worker:requirements_test also needed to be increased to a
long test time, as it was now timing out on `moderate`.

 * This change introduces two visibility patches: one for TFF
and one for TF that is requires due to TF dependencies within TFF. The
TF patch will be required going forward. In parallel to getting this
change reviewed, I will work to update the visibility inside the TFF
codebase so that the other patch is not necessary.

Change-Id: I1dcf5b1174b0e704d2bcb58f71c9df70d4753984