Update federated-compute dependency to a newer version.
This will pull in the recent updates to the converters
library in the federated-compute project that are needed
for writing the C++ TFF PipelineTransform service. A prior attempt
inadvertently pulled in the version with the converters, but not with
the subsequent visibility change.

This update also pulls in the clean up of deprecated fields from the
ConfidentialCompute APIs. This requires cleaning up the Python TFF
container code.

Change-Id: I92f7864f783a9ea9ee0f2b382c33c51178e4f839