Fix --config=asan build failures.
The version of TFF we're using declares but doesn't define
OwnedValueId::INVALID_ID, which causes linker errors in some contexts --
such as asan builds. This will fix the CI failures.

Also switch tensorflow-federated from a git_archive (requires git) to
a http_archive and make it reproducible by adding a hash. This makes it
consistent with other dependencies in the WORKSPACE file.

Change-Id: If6ff7bed5489fc392ce9e5f5ce13b377def56970