Set encrypted output node id in square_service.
This example service previously left the access_policy_node_id field
unchanged when producing encrypted outputs, which is incorrect.
Ideally, the destination node id would be provided by the ledger based
on what transform rule matched, but that's tied up with a future
project. In the interim, allow the node id to be passed in as
configuration and verified by the DataAccessPolicy.

The pipeline_transform crate also needed to be updated to support
setting the config properties claim.

Since this is a temporary solution in an example binary, there isn't as
much testing of error cases as we'd normally do. Please let me know if
you think that's a problem.

Bug: 287284320
Change-Id: I5bdfdcc887bde77908af7864a924572c33fa9299