Add "noubsan" tag to cpp_pipeline_transform_server_test target.
We are enabling sanitizers for the code/tests in the CFC repo. However,
this target is currently failing with
`relocation R_X86_64_PC32 out of range` linker errors.

Until we can figure out a solution, we are turning off the undefined
behavior sanitizer for the cpp_pipeline_transform_server_test target.

In order to have the sanitizer presubmit respect the tags, the
bazel_build.sh script was updated to respect tags turning off
sanitizers, both for building and testing.

Change-Id: I607c746d01388f0fd2fce60815bc891d256ff07b