Remove Python tff_worker and move C++ impl to containers/.
The Python tff_worker is no longer being used and is costly to maintain
(e.g. it requires a separate build process for the container). The
remaining partial C++ implementation can be moved to the `containers`
directory for consistency with other Oak Containers enclave apps.

Unfortunately, the Python and pypi deps in the WORKSPACE file can't
(yet?) be removed because they're needed by TensorFlow and by the
tff_worker's generate_test_computations tool.

Also rename *.pbtxt to *.txtpb since the latter is the canonical
extension.

Change-Id: Ice22a72b9033eaa230de0a45d99ab015686ad1b7