Describe how FCP attestation records relate to the binaries in this repo.
This change adds some initial documentation describing how attestation
verification records generated by the Federated Compute library can be
mapped to the actual ledger application that as attested to by the
evidence in those records, and to the data transformation applications
allowed by the data access policy in the verification record.

Note that these docs are expected to be expanded on in the coming
days/weeks.

Bug: 335273029
Change-Id: If60d1d3bb46b7197548bd4493a230f2dc7209fde