Return non-retriable error for data access denied.
RESOURCE_EXHAUSTED usually indicates that quota is exhausted and thus
will be retried automatically by some client libraries. But once the
access budget for some data is exhausted, retrying will never succeed.
Thus a PERMISSION_DENIED error should be used to reflect that access to
the data will not be granted even on retry.

Change-Id: Ib9a6a6135d57a0628cb6df051103623c9fc24895