Make the aggcore container use `std::transform` to accumulate records.
Previously we were creating a thread for each record, which could cause
a huge spike in memory usage for requests with many Records. We noticed
the Transform call throwing an exception when running a benchmark of
100k clients, and we think this is the cause.

We'd really like to limit the number of threads to the number of cores
on the machine.

Using `transform` with a parallel execution policy is not guaranteed to
run in parallel or to run on a reasonable number of threads... but we
hope it does.

Benchmarks show improved wall time and similar CPU time:
http://shortn/_yTtr7AJoVR

Change-Id: Ia893130788964366be8c83c6b79a291b9e78e84d