Split the Ledger create_key method into produce_create_key_event and apply_create_key_event
This is a step towards allowing key creation to be replicated.
The create_key method is now implemented by procuding an event and promptly applying it.

Additionally this change relaxed the requirement that the time is monotonic because we can't
guarantee that in the replicated version of Ledger. Instead the max of the provided time
and the current time is used both when initially handling the CreateKeyRequest and when
applying the replicated event.

Change-Id: Ia6eac1a1e8ef5c9c594b5946a980e1883c26c541