First version of a tool explaining FCP AttestationVerificationRecord protos.
This tool uses the Oak-provided `oak_attestation_explain` library to
produce a human-readable summary of the attestation evidence and
reference values in these records, along for producing a human readable
explanation of the data access policy in the record.

This is really just a first pass, and we'll likely want to evolve this
in the future to increase its usefulness. Some known limitations right
now:
- The specific commits from which binaries at various layers of the
  evidence/reference values were built are not included in the output of
  the tool. The output contains links to sigstore.dev for each binary in
  the ledger attestation evidence, however, and those attestations
  contain the commit ID, links to the GitHub repo, as well as a summary
  of the build steps that were used to produce the binary with that
  hash. This should be enough for someone to reproduce those same steps
  and arrive at binaries with the same hash.
- The `ReferenceValues` summary is basically a JSON dump of the binary
  digest hashes for each layer specified by the reference values.
- The configuration property matchers in the `ApplicationMatcher`
  message of a `DataAccessPolicy` is just a rust-debug-formatted dump of
  the `StructMatcher` proto rather than something more human-readable.
- The lines we output aren't really wrapped to any specific line length,
  which might hamper readability a bit.

To test the tool I included a serialized `Evidence` proto from our
ledger application running in a real TEE.

Lastly, this change also improves our build scripts a bit by adding
`set -x` in various places, which helps improve visibility into the
commands the scripts invoke as they execute.

Change-Id: I0a1b950054495f36b6299e8c26b6b6667466c30b