Expand inspecting_attestation_records/README.md to cover the new explanation tool.
This expands the README.md file to show how the
`explain_fcp_attestation_record` tool can be used to describe an FCP
`AttestationVerificationRecord`. It also elaborates a little bit more on
the information found in the SLSA provenance records on sigstore.dev,
including showing how the SLSA record describes the git repo and commit,
as well as the build command with which the ledger binary was produced.

Change-Id: I337b9ecc0d2ca8da8754a37d664155e028f3167e