Make aggregation container resilient to bad inputs.
If an input cannot be decrypted or parsed but the CheckpointAggregator
is still guaranteed to be in an unmodified, valid state, this change
updates the PipelineTransform server to count the
input in num_ignored_inputs but continue aggregation.

Also adds a test to ensure that a valid result is produced when no
inputs are provided in the Transform call, which could be the case if no
inputs are authorized.

BUG: 339677404
Change-Id: I9c3463f50a275866a9a64b85af86705a12f68bc9