Add initial FedSQL container that implements the new ConfidentialTransform API.
Currently the container uses the checkpoint aggregator to implement the
aggregation step of confidential FedSQL. In a following CL, I plan to
add the per-client SQL query step to this container. Instead of
using two separate containers for confidential FedSQL, we will use this
single container to handle both the per-client and aggregation steps.

Add BlobDecryptor which forks RecordDecryptor and removes the nonce
tracking logic. Nonce tracking is instead handled in
SessionNonceTracker, since nonces are now tracked locally within a
session, and BlobDecryptor is shared across aggregation sessions. We
should remove RecordDecryptor once we transition the existing containers
to the new ConfidentialTransform API.

Concatenate the session-level nonce and counter rather than using
absl::Hash because absl::Hash is not deterministic across processes, and
we need the has of the session-level nonce and counter to match across
the client and server.

Features that will be implemented in separate CLs:

- Use threads to accumulate blobs concurrently

- Handle chunking (blobs that span multiple WriteRequests)

- Tracking available memory within each aggregation session. Currently,
the container simply limits the number of ongoing aggregaton sessions.

Change-Id: Ifffef7dd8ff9b8cbe9321c7cf45c53b2d50801f5