Fail aggregation if no inputs have been aggregated.
Workflows that haven't aggregated any inputs should be marked as failed
rather than succeeded as this likely indicates some configuration issue
or bug.

If no inputs are aggregated but the operation is SERIALIZE rather than
REPORT, the operation should succeed as there is still an opportunity
for the empty input to be aggregated with other partial aggregates and
produce avalid result.

BUG: 354248240
Change-Id: I341370265d0783d3ed65b9058014391b0f9f1764