Move test tag filters to .bazelrc and add "no" prefix.
The previous "asan" filter doesn't make it clear that the
target shouldn't run in that configuration, and in practice
people used "noasan" instead.

Moving the tags to .bazelrc ensures that they're applied to
all builds, regardless of how they're invoked.

This also removes the build_tag_filters since it's useful to
verify that a target compiles during CI builds, even if it's
not possible to run the test with asan.

Bug: 374131426
Change-Id: I1f0b49a8e9b7b855470ba8000797092887c18f67