Update to the new EndorsementReferenceValue fields.
The new fields allow claim verification and support key
rotation -- and the old fields are in the process of being
removed. A test ensures that the updated reference values
continue to accept the current ledger endorsements.

Also update to the latest Oak version (and TCP version for
compatibility) to guarantee that the attestation verification
library fully supports the new reference value fields.

Bug: 379268663
Change-Id: Id9feb6cf95fb6d9e9548b74cf0b6b357966609b3