Initial implementation of the KMS storage layer.
The implementation is missing several important features, such
as preconditions for updates and entry expiration; these
features will be added in subsequent changes.

In addition to adding the Storage struct and corresponding
proto messages, this change adds support for building protos
with `rust_prost_proto` instead of `cargo_build_script`; this
especially simplifies the management of proto dependencies.

Bug: 393146003
Change-Id: Ibcd8fe038ecc7356e93475763351bd2e745d73ee