Storage support for expiring entries.
Expiration will be used to support automatically removing old encryption
keys (or other time-bound state) -- without needing to explicitly
enumerate and remove entries with ReadRequests and UpdateRequests.

Also update existing tests to only match against the fields relevant for
the behavior under test; this will minimize the number of tests that
need to be updated as new functionality is added.

Bug: 393146003
Change-Id: I64b8b5bffa70eec8a710a3d8d5a3e80213b055eb