Implement a client for sending StorageRequests.
This client handles multiplexing multiple read and update requests in an
encrypted Oak session over a streaming gRPC connection. This gRPC
connection will be to a server running outside of the TEE; that server
will handle routing the messages to the TCP leader.

Connection errors are retried, so callers of the StorageClient can
perform read and update calls without worrying about the underlying Oak
session.

Each time a new session is established, an initializing update operation
is performed just in case this is the first connection to the TCP
cluster (or the TCP job was restarted). In almost all cases, this will
result in a FAILED_PRECONDITION error because initialization will
already have occurred -- but even in those cases, the initialization
request ensures that the connection to the TCP leader is operational.

Bug: 393146003
Change-Id: Id863413046cc6635ed2c5ad8c2d6379307550a52