Re-enable a standard RBE image, this time with a clang sysroot.
Since the clang toolchain didn't previously have a sysroot, it relied on
libc++ being available in the build image -- which stopped being true
when we switched to the standard debian11 image.

This is fixed by specifying a sysroot; we use a recent one from the
Chromium project. This sysroot probably isn't reproducible, but that's
okay so long as it's not used for release builds.

This change re-applies commit 47edc505f399cb819b23f460a6b016510aaa01f1.

Change-Id: I97fa0355760fb49af6b430758f864a1dbc4446b3