Update to bazel 7 and remove now-unnecessary rules_pkg patch.
The rules_pkg patch is no longer necessary if we set
`--@rules_python//python/config_settings:bootstrap_impl=script`; this
requires rules_python >= 0.33.0.

rules_apple and apple_support also needed to be updated to work with
bazel 7. We should be able to switch back to the TensorFlow-provided
versions with TF 2.19.0.

Bug: 401562000
Change-Id: I5726035fcaf9503c5ed8a58b9ded696eb8061c85