Switch back to `InstanceSigner`.
`InstanceSessionBinder` is not ready for production use yet, so we
switch to the recently re-added `oak_sdk_containers::InstanceSigner`.
Also switch to the `*_ref` `SessionConfig` methods to avoid extra clones
and obsolete CloneableAttester and CloneableEndorser.

This unfortunately requires plumbing changes that will need to be
reverted when switching back to `InstanceSessionBinder`, but I haven't
found a simpler alternative.

Also update to the TCP revision that contains the same fix.

Change-Id: I4854a15542d90a0076e1b20cf15d544f499bb70d