Add ability to execute a simple hardcoded federated program from C++.
This cl introduces two layers of pybind:
- C++ confidential_transform_server <-> python program_runner
- python execution_context <-> C++ computation_runner

Finding versions of the pip dependencies needed by federated_language
and org_tensorflow_federated that didn't result in errors when used
alongside these two layers of pybind was tricky. It appears that
dm-tree==0.1.7 must be used, which forces some downstream changes to
the versions of other dependencies. It did not seem possible to find a
compatible version of dp_accounting so a patch was introduced to remove
this dependency from TFF. Once we no longer need to depend on TF, the
pip dependency situation should become easier to manage.

This cl is meant to establish the initial code needed for executing
a python federated program from C++ and prove that there is a suitable
set of dependencies that works. It uses a hardcoded program and
hardcoded DataReadWrite service port for now. Also, only non-distributed
TF execution is supported in the execution stack for now. A basic
ReleaseManager allows results produced within the federated program to
be exported to untrusted space.

In follow-up cls, the program and DataReadWrite service port will become
customizable and the execution context/stack will become more flexible
(supporting distributed execution and XLA).

Change-Id: I918eff423b0664d49dc975dc8d1b9e8771863f8f