Specify `--remote_download_minimal` during CI builds
Since CI builds only care about the outcome of the build/test steps,
there's no need to download build outputs to the local machine.

Also pass `--keep_going` to postsubmit builds so that the full set of
failures is clearer. We omit this from presubmit failures so that errors
are surfaced sooner.

Bug: 408057656
Change-Id: Ibc829036679e532bb14be92d9ccfdea294d9813f