Add not_before and expiration_time to signing key endorsements
`not_before` is set to the time when the invocation was registered;
`expiration_time` is set to the time when it expires. This information
isn't strictly necessary for consumers of the signing key endorsement,
but it increases consistency with other CWTs the KMS produces and might
be useful for debugging.

Bug: 398874186
Change-Id: Ie585f5e893b1ec2dfd4e417df45c129ffb45da2e