Allow multiple release tokens per logical pipeline
Multiple release tokens are allowed if their state changes can be
combined into a single larger change (e.g., A->B, B->C => A->C). To do
this, we determine if the directed multigraph defined by the state
transitions contains an Eulerian trail:
https://en.wikipedia.org/wiki/Eulerian_path.

It's not clear whether existing pipelines will require multiple release
tokens, but it's part of the approved API, would be useful in some
circumstances, and the added complexity is well-encapsulated. We can
always revert this change later if it proves unnecessary.

Bug: 398874186
Change-Id: If100a7224a36cd08ca5c608f9cb713f9e88230bc