Implementation of RangeTracker for tracking intermediate budgets
This class is responsible for tracking ranges of blobs in
an arbitrary number of buckets, each bucket identified with
a string key. Initially these buckets will be matched to
encryption key IDs.

Additionally, this change improves proto message matcher
to allow it to optionally ignore the order of repeated
fields.

Change-Id: I997faca1c58920f2aa2a98402ae40564ba72f6e1