Log TCP log messages to default log sink
TCP's slog-based log messages will now be logged to the same log sink as
log-based log messages, allowing the two to be managed consistently.
This also addresses the TCP logs being formatted for a terminal -- and
therefore including control codes that do not render well in a log file.

Fixed: 415795230
Change-Id: I86a258fd4c1023ae055034c38a830643c16f45c0