Add the noise client session for program executor TEE.
This helper class is used to create a client session on the program executor. This class will be used by the client side to delegate computation to the trusted worker. It handles the noise handshake to create an open session, the encryption and decryption of the messages, and call ComputationDelegation outgoing service to delegate the computation request to a specified worker.

Change-Id: I371b4f96b6ed002c5370a1b39c5ba62f1e76b964