Use the replicated clock for KMS-KMS attestation verification
By default, TCP uses a clock based on the time provided by the host in
the TCP request. While TCP ensures this value is non-decreasing, it
could still be different from the replicated timestamp -- and could be
different on different nodes in the cluster. Using a clock based on the
replicated timestamp ensures that the clock is consistent across all
nodes and that the same time is used for attestation verification as is
used for key TTLs.

When a node is first joining the cluster, it won't have loaded a
snapshot yet and will therefore not have a timestamp. In this specific
case, the current system time is used so that the node can successfully
verify the leader's endorsements. Once a snapshot is loaded, the
replicated time is used instead.

Fixed: 399699168
Change-Id: I8595f6fbc4c7e585526ffe8983dbf0caf2188231