Build a reproducible clang sysroot
This sysroot will allow the main CFC workspace to switch from an old
version of GCC to a more recent version of clang. The sysroot can be
re-built on demand using the "Release sysroot" workflow.

We introduce a new bazel workspace for the sysroot building because
`rules_distroless` only supports bzlmod.

Bug: 425785491
Change-Id: I0ec4997c0e04b4145957514803f2a33751e3076a