Fix the "Release sysroot" workflow
Since releases require a tag and it's standard to use that tag creation
to trigger a release, switch the workflow trigger to `sysroot-*` tag
creation. Also switch to a maintained release action to avoid warnings
about using deprecated APIs. This action is easier to use and allows the
release to not be marked as the latest, avoiding potential future
conflicts with non-sysroot releases.

Bug: 425785491
Change-Id: I474c27507cebc96fdb19f3991f5ec7b4720781b9