Use hermetic CUDA for minimum_rules_cuda container
This change also adds a sysroot for the clang compiler to make build
fully hermetic. With this change, we no longer rely on the dockerfile to
setup a build environment.

We also change the base layer from a Nvidia/CUDA docker image to a
distroless debian 12 image. We are statically linking most of the CUDA
dependencies, and the rest are included in the runfiles of the main
binary. Therefore, we don't need these CUDA deps in runtime anymore.

Bug: 373901407
Change-Id: I8eb2475361ff16c212f459babfe218bab3d90757