Add data uri resolution responsibilities to the execution context.
The TrustedAsyncContext is now responsible for resolving any Data
pointers found in an invoke arg by 1) making a ReadRequest to the
DataReadWrite service in untrusted space, and then 2) relying on a
helper function provided to the TrustedAsyncContext constructor to
parse (and potentially decrypt) the received ReadResponse into a
tff Value. In production code, this helper function will be a
callback to the confidential_transform_server.cc code to enable
decryption (this connection will be made in the next cl).

This cl also removes the python implementation of the fake
DataReadWrite service in favor of using the fake cpp implementation
everywhere. The fake service/server pybind binding code for the
fake DataReadWrite service and fake ComputationDelegation service
is also consolidated in one fake_service_bindings.cc file.

Change-Id: If892613d6e8ba17bd99d6abd07bf3a421ac22241