Allow a cpp callback for decryption during py-controlled data resolution.
The existing data_uri_resolver files are replaced with data_parser
files that provide the specific ReadResponse -> tff.Value
decryption and parsing logic we will need when the python execution
context is responsible for triggering data pointer resolution.
Since the BlobDecryptor is a cpp object and since the FC checkpoint
parser code is in cpp (client uploads will be FC checkpoints), a
helper function is provided in the confidential_transform_server.cc
FinalizeSession implemention to the python program runner so that
the execution context can make a callback to cpp for decryption and
parsing.

In order to avoid a pybind error involving TF proto redefinition,
it is important to avoid pulling in python proto dependencies in
both the cpp and python code that is connected via pybind within the
same process. Since the python code will import tensorflow, we do
not use the existing conversion functions from AggCore tensor ->
TF tensor -> federated_language Array in the cpp parsing logic.
Instead a new conversion helper function that goes directly from
AggCore tensor -> federated_language Array is used.

Change-Id: I604b2befd8147c4e2aeb73a1c1a9597fe8d2a980