Add integration test using min_sep_data_source.
This requires also updating the program_runner.py interface
to accept more information about the client ids and client
data dir. This information is now provided to the program
itself via an instance of the ProgramInputProvider.

A patch is added for the min_sep_data_source in FCP to
temporarily allow the computation type to be overwritten to
something other than str.

Change-Id: I066116478c417a6286a09a2c9acebdfda32bd4a1