Adjust pip versions to be more similar to those used by other repos.
Because we've recently replaced the second layer of pybind in the
program_executor_tee container with grpc, we no longer need to use
dm-tree==0.1.7 and can instead use dm-tree==0.1.8, which is what
federated_language, org_tensorflow_federated, federated_language_jax,
and federated-compute all expect. With this change we can also
include the dp-accounting pip package and remove the dp_accounting
patch we were using for org_tensorflow_federated.

Change-Id: I879573b646ac391d1da4c4488ab878d71494015f